[Delete] #401 - 불필요한 DTO, Service 파일 삭제
